Implement a base Storage plugin
