+++ packages/arbor-plugins/src/LocalStorage.ts
@@ -1,13 +1,13 @@
-import { Arbor, Plugin } from "@arborjs/store"
-
-import debounce from "./debounce"
+import Storage from "./Storage"
 
 /**
  * Describes a LocaStorage configuration object.
  */
 export interface Config<T extends object> {
+  /**
+   * The local storage key to reference the store data.
+   */
   key: string
-
   /**
    * Amount in milliseconds to debouce by local storage updates.
    *
@@ -15,7 +15,6 @@ export interface Config<T extends object> {
    * back to local storage.
    */
   debounceBy?: number
-
   /**
    * Serializes some data structure into its string representation to be
    * stored in the browser's local storage.
@@ -24,7 +23,6 @@ export interface Config<T extends object> {
    * @returns the string representation of the data.
    */
   serialize?: (data: T) => string
-
   /**
    * Deserializes the data retrieved from local storage.
    *
@@ -40,75 +38,41 @@ export interface Config<T extends object> {
  * This provides LocalStorage-based persistence to Arbor state trees allowing
  * application state to be preserved across browser refreshes/sessions.
  */
-export default class LocalStorage<T extends object> implements Plugin<T> {
-  private deboucedUpdate: (data: T) => void
-
+export default class LocalStorage<T extends object> extends Storage<T> {
   /**
    * Creates a new instance of the LocalStorage backend.
    *
    * @param key the key to be used in LocalStorage to reference the serialized state.
    */
-  constructor(readonly config: Config<T>) {
-    this.deboucedUpdate = debounce((data: T) => {
-      const serialize =
-        config.serialize || (JSON.stringify as typeof config.serialize)
+  constructor(config: Config<T>) {
+    super({
+      debounceBy: config.debounceBy,
 
-      window.localStorage.setItem(this.config.key, serialize(data))
-    }, config.debounceBy)
-  }
+      load: () => {
+        return new Promise<T>((resolve, reject) => {
+          try {
+            const data = window.localStorage.getItem(config.key) || "null"
+            const deserialize =
+              config.deserialize || (JSON.parse as typeof config.deserialize)
 
-  /**
-   * Runs the plugin's configuration/initialization logic.
-   *
-   * @param store the Arbor store to apply the plugin to.
-   * @returns a Promise that resolves when configuration logic completes.
-   */
-  async configure(store: Arbor<T>) {
-    const data = await this.load()
+            resolve(deserialize(data))
+          } catch (e) {
+            reject(e)
+          }
+        })
+      },
 
-    if (data && typeof data === "object") {
-      store.setState(data)
-    }
-
-    store.subscribe(({ state }) => {
-      void this.update(state.current)
-    })
-  }
-
-  /**
-   * Loads the persisted data from window.localStorage.
-   *
-   * @returns a promise that resolves to the persisted data if successful.
-   */
-  async load(): Promise<T> {
-    return new Promise<T>((resolve, reject) => {
-      try {
-        const data = window.localStorage.getItem(this.config.key) || "null"
-        const deserialize =
-          this.config.deserialize ||
-          (JSON.parse as typeof this.config.deserialize)
-
-        resolve(deserialize(data))
-      } catch (e) {
-        reject(e)
-      }
-    })
-  }
-
-  /**
-   * Updates the LocalStorage with the given data.
-   *
-   * @param data the data to be persisted to window.localStorage.
-   * @returns a Promise that resolves to void when the operation completes.
-   */
-  async update(data: T): Promise<void> {
-    return new Promise((resolve, reject) => {
-      try {
-        this.deboucedUpdate(data)
-        resolve()
-      } catch (e) {
-        reject(e)
-      }
+      update: (data: T): Promise<void> => {
+        return new Promise((resolve, reject) => {
+          try {
+            const serialize = config.serialize || JSON.stringify
+            window.localStorage.setItem(config.key, serialize(data))
+            resolve()
+          } catch (e) {
+            reject(e)
+          }
+        })
+      },
     })
   }
 }
